app/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
httpx==0.27.2
python-dotenv==1.0.1
uvicorn[standard]==0.30.5

